Improve admin dashboard & update readme (#31)
* Reduce duplication and improve modularity in welcome pages

* Reduce duplication and improve modularity in dashboard pages

* Add js scrollTo function and connect event navbar with db

* Add routing and controller & improve modularity

* Rename footer icon

* Add scroll after redirect feature

* Dynamic html title on each page

* Reduce duplication and improve modularity on about pages

* Add messages db and its behaviour and add flash messages on form submission

* Make validation for contact form and change error response to b.indo

* Add events data dependency for layouts.guest

* Improve Restful controller design

* Add access to admin panel

* Store payment receipt

* Update seeder

* Add mail notification for payment status

* Add free event

* Fix test error

* Add authorization

* Add notification for number of pending status

* Add verification menu design

* Add pagination adn improve responsivenss on verify menu

* Add confirmation modal

* Change timezone & update pivot table with ajax

* Make alert on accept & reject payment receipt

* Add loading bar

* Redesign email

* Implement register status card

* Payment feature finishing

* Improve pagination design

* Add edit announcements feature

* Improve views

* Add stats card

* Finish admin panel

* Change button padding

* Commented test_new_users_can_register

* Update readme

* Fix test

* Export to excell based on events

* Add query db

* Improve authorization

+++ app/Http/Controllers/MessagesController.php
@@ -10,6 +10,7 @@ class MessagesController extends Controller
 {
     public function index()
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         $messages = Messages::all();
 
         return view('dashboard.admin.message.index', compact('messages'));
@@ -17,6 +18,7 @@ class MessagesController extends Controller
 
     public function show(Messages $message)
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         return view('dashboard.admin.message.show', compact('message'));
     }
 
@@ -26,6 +28,7 @@ class MessagesController extends Controller
 
     public function update(Request $request, Messages $message)
     {
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
         $message->update($request->validate([
             'status' => [
                 'required',
@@ -56,8 +59,14 @@ class MessagesController extends Controller
 
     public function destroy(Messages $message)
     {
-        $message->delete();
+        $this->authorize('interactAsMaster'); // If false, it'll display 403
 
-        return redirect()->route('message.index');
+        if ($message->status == 'Sudah diproses') {
+            $message->delete();
+            return redirect()->route('message.index');
+        } else {
+            return redirect()->route('message.index')
+                ->with('error', 'Pesan belum dapat dihapus!');
+        }
     }
 }
